a) Corrected treatment of Line-In gain (now from -34 to +12 dB), b) better organization of the "Slider" area

+++ radio.h
@@ -154,7 +154,7 @@ extern int display_toolbar;
 extern double mic_gain;
 
 extern int mic_linein;
-extern int linein_gain;
+extern double linein_gain;
 extern int mic_boost;
 extern int mic_bias_enabled;
 extern int mic_ptt_enabled;
